+++ CHANGELOG.md
@@ -5,6 +5,12 @@ file is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and
 this project adheres to [Semantic
 Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- `enum DataSourceKey`, which is used in fetching a data source from the server.
+
 ## [0.8.0] - 2024-10-11
 
 ### Added
@@ -250,6 +256,7 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 - `client::handshake` implements the application-level handshake process for the
   client after a QUIC connection is established.
 
+[Unreleased]: https://github.com/petabi/review-protocol/compare/0.8.0...main
 [0.8.0]: https://github.com/petabi/review-protocol/compare/0.7.0...0.8.0
 [0.7.0]: https://github.com/petabi/review-protocol/compare/0.6.0...0.7.0
 [0.6.0]: https://github.com/petabi/review-protocol/compare/0.5.0...0.6.0